Race status result type
