+++ packages/common/src/store/buy-crypto/sagas.ts
@@ -60,24 +60,7 @@ import { initializeStripeModal } from 'store/ui/stripe-modal/slice'
 
 import { BuyAudioStage, OnRampProvider } from '..'
 
-import { BuyCryptoConfig } from './types'
-
-export enum BuyCryptoErrorCode {
-  BAD_AMOUNT = 'BadAmount',
-  BAD_TOKEN = 'BadToken',
-  BAD_PROVIDER = 'BadProvider',
-  BAD_FEE_PAYER = 'BadFeePayer',
-  SWAP_ERROR = 'SwapError',
-  ON_RAMP_ERROR = 'OnRampError',
-  UNKNOWN = 'UnknownError'
-}
-
-class BuyCryptoError extends Error {
-  name = 'BuyCryptoError'
-  constructor(public code: BuyCryptoErrorCode, message: string) {
-    super(`${code}: ${message}`)
-  }
-}
+import { BuyCryptoConfig, BuyCryptoError, BuyCryptoErrorCode } from './types'
 
 function* getBuyAudioRemoteConfig() {
   // Default slippage tolerance, in percentage basis points
