+++ packages/common/src/store/purchase-content/sagas.ts
@@ -20,6 +20,7 @@ import {
   buyCryptoSucceeded,
   buyCryptoViaSol
 } from 'store/buy-crypto/slice'
+import { BuyCryptoError } from 'store/buy-crypto/types'
 import {
   buyUSDCFlowFailed,
   buyUSDCFlowSucceeded,
@@ -52,6 +53,18 @@ import {
 import { ContentType, PurchaseContentError, PurchaseErrorCode } from './types'
 
 const { getUserId } = accountSelectors
+
+type RaceStatusResult = {
+  succeeded?:
+    | ReturnType<typeof buyUSDCFlowSucceeded>
+    | ReturnType<typeof buyCryptoSucceeded>
+  failed?:
+    | ReturnType<typeof buyUSDCFlowFailed>
+    | ReturnType<typeof buyCryptoFailed>
+  canceled?:
+    | ReturnType<typeof onrampCanceled>
+    | ReturnType<typeof buyCryptoCanceled>
+}
 
 type GetPurchaseConfigArgs = {
   contentId: ID
@@ -227,8 +240,7 @@ function* doStartPurchaseContentFlow({
         .div(BN_USDC_CENT_WEI)
         .toNumber()
       yield* put(buyUSDC())
-      let result: { succeeded?: any; failed?: any; canceled?: any } | null =
-        null
+      let result: RaceStatusResult | null = null
       if (isBuyUSDCViaSolEnabled) {
         yield* put(
           buyCryptoViaSol({
@@ -318,7 +330,9 @@ function* doStartPurchaseContentFlow({
     // If we get a known error, pipe it through directly. Otherwise make sure we
     // have a properly contstructed error to put into the slice.
     const error =
-      e instanceof PurchaseContentError || e instanceof BuyUSDCError
+      e instanceof PurchaseContentError ||
+      e instanceof BuyUSDCError ||
+      e instanceof BuyCryptoError
         ? e
         : new PurchaseContentError(PurchaseErrorCode.Unknown, `${e}`)
     yield* call(reportToSentry, {
