+++ packages/common/src/store/purchase-content/types.ts
@@ -1,3 +1,5 @@
+import { BuyCryptoErrorCode } from 'store/buy-crypto/types'
+
 import { BuyUSDCErrorCode } from '../buy-usdc'
 
 export enum ContentType {
@@ -17,7 +19,10 @@ export enum PurchaseErrorCode {
   Unknown = 'Unknown'
 }
 
-export type PurchaseContentErrorCode = BuyUSDCErrorCode | PurchaseErrorCode
+export type PurchaseContentErrorCode =
+  | BuyUSDCErrorCode
+  | BuyCryptoErrorCode
+  | PurchaseErrorCode
 
 export class PurchaseContentError extends Error {
   constructor(public code: PurchaseContentErrorCode, message: string) {
